Export batcher as a transform instead of a Stream method
