+++ src/test/util.ts
@@ -7,6 +7,8 @@
  */
 
 import { defer } from "../lib/util";
+import { ReadableStream } from "../lib";
+import { expect } from "chai";
 export * from "../lib/util";
 
 /**
@@ -26,4 +28,34 @@ export function settle(promises: PromiseLike<any>[]): Promise<void> {
 		promise.then(check, check);
 	}
 	return result.promise;
+}
+
+export function readInto<T>(
+	stream: ReadableStream<T>,
+	into: T[]
+): Promise<void> {
+	return new Promise<void>((resolve, reject) => {
+		stream.forEach(
+			function (this: any, value: T): void {
+				expect(this).to.equal(undefined);
+				into.push(value);
+			},
+			function (this: any, err?: Error): void {
+				expect(this).to.equal(undefined);
+				if (err) {
+					reject(err);
+				} else {
+					resolve(undefined);
+				}
+			}
+		);
+	});
+}
+
+export function noop(): void {
+	/* empty */
+}
+
+export function identity<T>(arg: T): T {
+	return arg;
 }
